[CI] Fixed some warnings from the builds

+++ Source/Editor/UI/Windows/AssetsBrowser/src/AssetsBrowser.cpp
@@ -738,7 +738,7 @@ void AssetsBrowser::Draw()
         }
         ImGui::EndChild();
 
-        ImGui::Text("Selected: %d/%llu items", Selection.Size, Items.size());
+        ImGui::Text("Selected: %d/%llu items", Selection.Size, (unsigned long long)Items.size());
         ImGui::End();
     }
 }
