+++ Source/Editor/UI/Windows/MemoryWindow/src/MemoryWindow.cpp
@@ -137,7 +137,7 @@ void MemoryWindow::Draw()
                 ImGui::TableSetColumnIndex(0);
                 ImGui::TextUnformatted(tags[i]);
                 ImGui::TableSetColumnIndex(1);
-                ImGui::Text("%llu", bytes);
+                ImGui::Text("%llu", (unsigned long long)bytes);
                 ImGui::TableSetColumnIndex(2);
                 ImGui::Text("%.3f", bytes / (1024.0f * 1024.0f));
             }
@@ -153,7 +153,7 @@ void MemoryWindow::Draw()
             ImGui::TextColored(ImVec4(0.9f, 0.8f, 0.3f, 1.0f), "TOTAL");
 
             ImGui::TableSetColumnIndex(1);
-            ImGui::TextColored(ImVec4(0.9f, 0.8f, 0.3f, 1.0f), "%llu", totalBytes);
+            ImGui::TextColored(ImVec4(0.9f, 0.8f, 0.3f, 1.0f), "%llu", (unsigned long long)totalBytes);
 
             ImGui::TableSetColumnIndex(2);
             ImGui::TextColored(ImVec4(0.9f, 0.8f, 0.3f, 1.0f), "%.3f", totalBytes / (1024.0f * 1024.0f));
